refactor event code and added timer heartbeat with sequence
I added the Opcodes to be able to handle other events that come back from discord.  I refactored the structure of the main loop to add sequencing for all of the DiscordOpcodes.DISPATCH events that are fired.  I noticed the heartbeat wasn't doing the jitter pause on startup and wasn't sending the sequence number for resumes, so I added the jitter and sequence on a actual timer so there's no need to keep recreating the timer.